+++ src/main/java/cse213/final_project/sadika/helpdesksupporter.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 
 public class helpdesksupporter extends User {
 
-    public helpdesksupporter(String address, LocalDate dob, String email, int id, String name, String password, String pN, String type) {
+    public helpdesksupporter(String address, LocalDate dob, String email, Long id, String name, String password, String pN, String type) {
         super(address, dob, email, id, name, password, pN, type);
     }
 }
