+++ src/main/java/cse213/final_project/sadika/registrationOfficer.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 public class registrationOfficer extends User {
 
 
-    public registrationOfficer(String address, LocalDate dob, String email, int id, String name, String password, String pN, String type) {
+    public registrationOfficer(String address, LocalDate dob, String email, Long id, String name, String password, String pN, String type) {
         super(address, dob, email, id, name, password, pN, type);
     }
 }
